Introduce negative cache
Reduce latency when repeatedly looking up non-existing files or directories (when cache is enabled).

This change adds negative metadata cache entries: whenever a lookup fails because an object does not exist, we cache a “negative” entry with the same TTL as for successful lookups and use it to reply to subsequent kernel requests for the same name.

The negative entries are maintained separately from the inode tree using the new `NegativeCache` type, which enforces an upper limit to the number of entries and handles their expiration.

Signed-off-by: Alessandro Passaro <alexpax@amazon.co.uk>